+++ dark-matter-data/src/main/java/me/melontini/dark_matter/api/data/loading/ReloaderType.java
@@ -1,5 +1,9 @@
 package me.melontini.dark_matter.api.data.loading;
 
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.experimental.Accessors;
 import net.fabricmc.fabric.api.resource.ResourceReloadListenerKeys;
 import net.minecraft.loot.LootManager;
 import net.minecraft.recipe.RecipeManager;
@@ -8,8 +12,13 @@ import net.minecraft.resource.ResourceReloader;
 import net.minecraft.server.ServerAdvancementLoader;
 import net.minecraft.server.function.FunctionLoader;
 import net.minecraft.util.Identifier;
+import org.jetbrains.annotations.Contract;
+import org.jetbrains.annotations.NotNull;
 
-public record ReloaderType<T extends ResourceReloader>(Identifier identifier) {
+@Getter
+@RequiredArgsConstructor(access = AccessLevel.PRIVATE)
+@Accessors(fluent = true)
+public final class ReloaderType<T extends ResourceReloader> {
 
     public static final ReloaderType<TagManagerLoader> TAGS = create(ResourceReloadListenerKeys.TAGS);
     public static final ReloaderType<RecipeManager> RECIPES = create(ResourceReloadListenerKeys.RECIPES);
@@ -17,7 +26,10 @@ public record ReloaderType<T extends ResourceReloader>(Identifier identifier) {
     public static final ReloaderType<FunctionLoader> FUNCTIONS = create(ResourceReloadListenerKeys.FUNCTIONS);
     public static final ReloaderType<LootManager> LOOT_TABLES = create(ResourceReloadListenerKeys.LOOT_TABLES);
 
-    public static <T extends ResourceReloader> ReloaderType<T> create(Identifier identifier) {
+    private final Identifier identifier;
+
+    @Contract("_ -> new")
+    public static <T extends ResourceReloader> @NotNull ReloaderType<T> create(Identifier identifier) {
         return new ReloaderType<>(identifier);
     }
 }