+++ dark-matter-data/src/main/java/me/melontini/dark_matter/impl/data/codecs/package-info.java
@@ -1,0 +1,4 @@
+@ApiStatus.Internal
+package me.melontini.dark_matter.impl.data.codecs;
+
+import org.jetbrains.annotations.ApiStatus;